+++ resources/js/admin/main.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+import { createRoot } from "react-dom/client";
+import Router from "./Router.jsx";
+
+createRoot(document.getElementById("app-admin-content")).render(<Router />);